Distribute floating ips on multiple nodes
This commit will distribute floating ips on multiple nodes. It currently
only does this with unassigned floating ips or floating ips that are on
unhealthy nodes. (Ips will not be distributed when additional nodes will
be added to the cluster)

This is done by getting the internal/external ip from all nodes that
are healthy and finding the corresponding hetzner server, just like the
controller node found its server before.
After that all floating on no or unhealthy nodes will be (re)assigned to
all running nodes, starting with the nodes that has the fewest floating
ips assigned.