Feature/ips_on_multiple_nodes (#39)
* Only COPY go.mod and go.sum before downloading deps

This makes docker builds signinficantly faster since docker does not
need to always fetch all go dependencies.

* Add a conf option to select floating ips by label

* Distribute floating ips on multiple nodes

This commit will distribute floating ips on multiple nodes. It currently
only does this with unassigned floating ips or floating ips that are on
unhealthy nodes. (Ips will not be distributed when additional nodes will
be added to the cluster)

This is done by getting the internal/external ip from all nodes that
are healthy and finding the corresponding hetzner server, just like the
controller node found its server before.
After that all floating on no or unhealthy nodes will be (re)assigned to
all running nodes, starting with the nodes that has the fewest floating
ips assigned.

* Return error if no ip addresses for any nodes are found

* Update docs with new ENV label selector

Explain new configuration value FloatingIPsLabelSelector and update docs
for HcloudFloatingIPs

* Remove old comments in controller.go

* Add label selector for nodes as well

* Move getFloatingIps functions to hcloud.go

* Add better error messages

* Remove wrong flag name

* Make hcloud.go more readably by creating new method for inner for loop

* Return nil instead of struct pointer when returning error

* Change sorting servers once to always retreive minimum server

* Add dummy floating ip to server after it got a real floating ip for searchmin purposes

* Add pod label selector option

Add a option to specify a pod label in case the default label needs to
be switched

* Add option to find nodes by deployment pods

This commit adds an option to find the hetzner server through the
deployed fip-controller pods.

This is done by searching for a specific pod label that can be created
with the deployment (label can be configured).

If no pods are found this way it will try to get the servers the old way
by finding the nodes

* Remove bug to for finding server with privatenet

previously the name of the privatenet was returned when a correct
privatenet was found.
Because this fields is not filled in when querying the servers (only the
id is), the method now returns just a boolean.

* update configuration docs

* Fixed code according to go vet and go lint

Fixed some formatting errors and especially wrong and missing godoc
formatted commentary

* Remove typo in docs/configuration.md

Replace defailt with default

* Fix typo in commentary from hcloud.go

Change give to given

* Move code from function searchPrivateNet to searchServerForIP

* Change function findServerByID to hasServerByID

Since the behaviour of the function changed at some point, the function
name as well as the comment should be updated as well.

* Unify style of function comments

* Return an error if no healthy nodes can be found

* Clarify why Set method is needed

Before the comments did not specify why Set is needed for the
nodeaddresstype

* Clarify pod_label_selector in documentation

* Change default pod label selector to all own labels

The default pod label selector will now use all its own key value
pairs, when no custom label selector is specified.

* Add rbac permissions for pod get and list

* Add debug output with created label selector

* Add checks for emtpy or nil lists in controller.go